+++ rma/tasks/filters.py
@@ -1,14 +1,17 @@
 import re
 import json
 
+from rma.utils import get_logger
 from rma.tasks.executor import Executor
 from rma.constants import ED_KWDS_PATTERN
+
+logger = get_logger(__name__)
 
 
 class FilterPostsScoreAboveMean(Executor):
     def __init__(self, mean, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.mean = int(mean)
+        self.mean = float(mean)
 
     def handle_msg(self, msg):
         if int(msg["score"]) >= self.mean:
@@ -21,6 +24,7 @@ class FilterPostsScoreAboveMean(Executor):
 class FilterEdComment(Executor):
     def handle_msg(self, msg):
         if re.search(ED_KWDS_PATTERN, msg["body"].lower()) is not None:
+            del msg["body"]  # TODO: remove from here
             self.task_out.send(json.dumps(msg).encode())
 
     def final_stmt(self):
@@ -29,8 +33,11 @@ class FilterEdComment(Executor):
 
 class FilterNanSentiment(Executor):
     def handle_msg(self, msg):
-        if msg["sentiment"] is not None:
+        try:
+            float(msg["sentiment"])
             self.task_out.send(json.dumps(msg).encode())
+        except:  # pylint: disable=bare-except
+            pass
 
     def final_stmt(self):
         pass
@@ -38,7 +45,7 @@ class FilterNanSentiment(Executor):
 
 class FilterNullUrl(Executor):
     def handle_msg(self, msg):
-        if msg["url"] is not None:
+        if msg["url"] is not None and msg["url"] != "":
             self.task_out.send(json.dumps(msg).encode())
 
     def final_stmt(self):
@@ -51,7 +58,9 @@ class FilterUniqPosts(Executor):
         self.post_ids = set()
 
     def handle_msg(self, msg):
-        self.post_ids.add({"post_id": msg["post_id"]})
+        self.post_ids.add(msg["id"])
 
     def final_stmt(self):
-        self.task_out.send(json.dumps(list(self.post_ids)).encode())
+        for post_id in self.post_ids:
+            msg = {"id": post_id}
+            self.task_out.send(json.dumps(msg).encode())