+++ rma/tasks/transforms.py
@@ -11,9 +11,13 @@ logger = get_logger(__name__)
 
 class ExtractPostID(Executor):
     def handle_msg(self, msg):
-        msg["post_id"] = re.match(
+        msg_id = re.match(
             r"https://old.reddit.com/r/me_?irl/comments/([^/]+)/.*", msg["permalink"]
-        ).groups()[0]
+        ).groups()
+        if msg_id is None:
+            logger.error("Bad permalink %s", msg["permalink"])
+            return
+        msg["id"] = msg_id[0]
         del msg["permalink"]
         self.task_out.send(json.dumps(msg).encode())
 
@@ -27,13 +31,13 @@ class MeanSentiment(Executor):
         self.sentiment_scores = defaultdict(lambda: {"count": 0, "sum": 0})
 
     def handle_msg(self, msg):
-        post_id = msg["post_id"]
+        post_id = msg["id"]
         self.sentiment_scores[post_id]["count"] += 1
-        self.sentiment_scores[post_id]["sum"] += msg["sentiment"]
+        self.sentiment_scores[post_id]["sum"] += float(msg["sentiment"])
 
     def final_stmt(self):
         for k, v in self.sentiment_scores.items():
-            msg = {"post_id": k, "mean_sentiment": v["sum"] / v["count"]}
+            msg = {"id": k, "mean_sentiment": v["sum"] / v["count"]}
             self.task_out.send(json.dumps(msg).encode())
 
 
@@ -47,6 +51,7 @@ class PostsScoreMean(Executor):
 
     def final_stmt(self):
         retval = sum(self.posts_scores) / len(self.posts_scores)
+        # TODO: send a dict
         self.task_out.send(json.dumps(retval).encode())